+++ src/Task.js
@@ -3,7 +3,16 @@ id = 1;
 
 const noop = () => {}; // eslint-disable-line no-empty-function
 
+/**
+ * @class Task
+ */
 class Task {
+    /**
+     * Tasks instances should only be created by Queue instance. Do not use directly.
+     * @param {Queue} queue - queue instance
+     * @param {function} fn - task function
+     * @param {function} check - function that verifies if task can be started
+     */
     constructor(queue, fn, check) {
         this._queue = queue;
         this._fn = fn;
@@ -39,20 +48,35 @@ class Task {
         return Promise.resolve();
     }
 
+    /**
+     * Checks if task was requested to be cancelled.
+     * @returns {boolean} - true if task was cancelled, false otherwise
+     */
     isCancelled() {
         return this._cancelled;
     }
 
+    /**
+     * Requests task cancellation.
+     */
     cancel() {
-        if (!this._cancelled) {
-            this._cancelled = true;
-            this._cancelReject(new Error("Task cancelled"));
-            if (!this._started) {
-                this.remove();
-            }
+        if (this._cancelled) {
+            return;
+        }
+        this._cancelled = true;
+        this._cancelReject(new Error("Task cancelled"));
+        if (!this._started) {
+            this.remove();
         }
     }
 
+    /**
+     * Starts task.
+     * @param {boolean} force - force start immediately
+     * @alias start
+     * @returns {void|Promise}
+     * @throws Error - when task is already started or task belongs to queue that is destroyed
+     */
     run(force) {
         if (!force && !this._check()) {
             return;
@@ -69,6 +93,10 @@ class Task {
         return this._fn(this._isCancelled.bind(this), this._cancelPromise).then(this._resolve, this._reject);
     }
 
+    /**
+     * Removes task from list without cancelling it.
+     * @deprecated - use cancel instead
+     */
     remove() {
         this._queue.remove(this);
     }